% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCongestionPolicy.R
\name{CalculateCongestionPolicy}
\alias{CalculateCongestionPolicy}
\title{Function to calculate the amount of congestion after adjusting to policy.}
\usage{
CalculateCongestionPolicy(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateCongestionPolicy} calculates the amount of congestion after adjusting
for policy.
}
\details{
Auto, and light truck vmt, truck vmt, and bus vmt are allocated to freeways, arterials,
and other roadways. Truck and bus vmt are allocated based on mode-specific data,
and auto and light truck vmt are allocated based on a combination of factors
and a model that is sensitive to the relative supplies of freeway and arterial
lane miles.

System-wide ratios of vmt to lane miles for freeways and arterials
are used to allocate vmt to congestion levels using congestion levels defined by
the Texas Transportation Institute for the Urban Mobility Report. Each freeway and
arterial congestion level is associated with an average trip speed for conditions that
do and do not include ITS treatment for incident management on the roadway. Overall average
speeds by congestion level are calculated based on input assumptions about the degree of
incident management. Speed vs. fuel efficiency relationships for light vehicles, trucks,
and buses are used to adjust the fleet fuel efficiency averages computed for the region.
}
