% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateCongestionBase.R,
%   R/CalculateCongestionFuture.R, R/CalculateCongestionPolicy.R
\docType{data}
\name{CongModel_ls}
\alias{CongModel_ls}
\title{Congestion models and required parameters.}
\format{A list having 'Fwy' and 'Art' components. Each component has a
logistic model to indicate the level of congestion which are categorized
as NonePct, HvyPct, SevPct, and NonePct. This list also contains other
parameters that are used in the evaluation of aforementioned models.}
\source{
GreenSTEP version ?.? model.

GreenSTEP version ?.? model.

GreenSTEP version ?.? model.
}
\usage{
CongModel_ls

CongModel_ls

CongModel_ls
}
\description{
A list of components describing congestion models and various parameters
required by those models.

A list of components describing congestion models and various parameters
required by those models.

A list of components describing congestion models and various parameters
required by those models.
}
\keyword{datasets}
